ref: modularized home page code
