+++ dashboard-ui/package.json
@@ -17,7 +17,7 @@
     "@headlessui/react": "2.2.0",
     "@heroicons/react": "^2.2.0",
     "@hookform/resolvers": "^5.2.0",
-    "@kubetail/ui": "2.0.0-rc5",
+    "@kubetail/ui": "2.0.0-rc6",
     "@radix-ui/react-checkbox": "^1.3.2",
     "@radix-ui/react-dropdown-menu": "^2.1.15",
     "@radix-ui/react-label": "^2.1.7",
@@ -32,6 +32,7 @@
     "date-fns-tz": "^3.2.0",
     "distinct-colors": "^3.0.0",
     "fancy-ansi": "^0.1.3",
+    "fast-deep-equal": "^3.1.3",
     "graphql": "^16.11.0",
     "graphql-ws": "^6.0.5",
     "jotai": "^2.12.5",
